+++ src/Services/Interfaces/IUserService.cs
@@ -11,8 +11,8 @@ namespace Services.Interfaces
     {
         Task<UserDTO> Create(UserDTO userDTO);
         Task<UserDTO> Update(UserDTO userDTO);
-        Task Remove(long id);
-        Task<List<UserDTO>> Get(long id);
+        Task Remove(int id);
+        Task<List<UserDTO>> Get(int id);
         Task<List<UserDTO>> SearchByUsuario(string usuario);
         Task<List<UserDTO>> SearchByNome(string nome);
         Task<UserDTO> GetByUsuario(string usuario);
